requirements.txt
google-generativeai
python-dotenv
requests
pandas
openpyxl
tabulate
pandas 
openpyxl


# pip install -r requirements.txt